fix(clash): split clash_solid.rs test module into clash_solid_tests.rs
The new exact-boundary test pushed clash_solid.rs to 417 lines, breaching
the module-size ratchet's 400-line budget and failing CI (`Rust tests`,
`no_module_grows_past_its_ratchet_budget`). Following the repo convention
for test-heavy modules (rust/core/src/georef.rs / georef_tests.rs), split
the inline #[cfg(test)] mod out to a sibling clash_solid_tests.rs file,
which the ratchet excludes as a *_tests.rs file. No test logic changed;
clash_solid.rs is back to 248 lines.
